Доработка сборщика cmake под Solaris

+++ src/sys/chrono.cpp
@@ -1935,35 +1935,35 @@ uint64_t awh::Chrono::actual(const uint64_t date, const type_t value, const type
 									// Получаем начало дня недели
 									const uint64_t begin = this->begin(date, type);
 									// Определяем количество прошедших дней
-									result = static_cast <uint64_t> ((static_cast <uint64_t> (604800000) - static_cast <uint64_t> (::floor(date - begin))) / 86400000.L);
+									result = static_cast <uint64_t> ((static_cast <uint64_t> (604800000) - static_cast <uint64_t> (::floor(static_cast <long double> (date - begin)))) / 86400000.L);
 								} break;
 								// Если нам нужно получить количество оставшихся часов
 								case static_cast <uint8_t> (type_t::HOUR): {
 									// Получаем начало дня недели
 									const uint64_t begin = this->begin(date, type);
 									// Определяем количество прошедших часов
-									result = static_cast <uint64_t> ((static_cast <uint64_t> (604800000) - static_cast <uint64_t> (::floor(date - begin))) / 3600000.L);
+									result = static_cast <uint64_t> ((static_cast <uint64_t> (604800000) - static_cast <uint64_t> (::floor(static_cast <long double> (date - begin)))) / 3600000.L);
 								} break;
 								// Если нам нужно получить количество оставшихся минут
 								case static_cast <uint8_t> (type_t::MINUTES): {
 									// Получаем начало дня недели
 									const uint64_t begin = this->begin(date, type);
 									// Определяем количество прошедших минут
-									result = static_cast <uint64_t> ((static_cast <uint64_t> (604800000) - static_cast <uint64_t> (::floor(date - begin))) / 60000.);
+									result = static_cast <uint64_t> ((static_cast <uint64_t> (604800000) - static_cast <uint64_t> (::floor(static_cast <long double> (date - begin)))) / 60000.);
 								} break;
 								// Если нам нужно получить количество оставшихся секунд
 								case static_cast <uint8_t> (type_t::SECONDS): {
 									// Получаем начало дня недели
 									const uint64_t begin = this->begin(date, type);
 									// Определяем количество прошедших секунд
-									result = static_cast <uint64_t> ((static_cast <uint64_t> (604800000) - static_cast <uint64_t> (::floor(date - begin))) / 1000.);
+									result = static_cast <uint64_t> ((static_cast <uint64_t> (604800000) - static_cast <uint64_t> (::floor(static_cast <long double> (date - begin)))) / 1000.);
 								} break;
 								// Если нам нужно получить количество оставшихся миллисекунд
 								case static_cast <uint8_t> (type_t::MILLISECONDS): {
 									// Получаем начало дня недели
 									const uint64_t begin = this->begin(date, type);
 									// Определяем количество прошедших миллисекунд
-									result = static_cast <uint64_t> (static_cast <uint64_t> (604800000) - static_cast <uint64_t> (::floor(date - begin)));
+									result = static_cast <uint64_t> (static_cast <uint64_t> (604800000) - static_cast <uint64_t> (::floor(static_cast <long double> (date - begin))));
 								} break;
 								// Если нам нужно получить количество оставшихся микросекунд
 								case static_cast <uint8_t> (type_t::MICROSECONDS): {
@@ -1990,28 +1990,28 @@ uint64_t awh::Chrono::actual(const uint64_t date, const type_t value, const type
 									// Получаем начало дня недели
 									const uint64_t begin = this->begin(date, type);
 									// Определяем количество прошедших часов
-									result = static_cast <uint64_t> ((static_cast <uint64_t> (86400000) - static_cast <uint64_t> (::floor(date - begin))) / 3600000.L);
+									result = static_cast <uint64_t> ((static_cast <uint64_t> (86400000) - static_cast <uint64_t> (::floor(static_cast <long double> (date - begin)))) / 3600000.L);
 								} break;
 								// Если нам нужно получить количество оставшихся минут
 								case static_cast <uint8_t> (type_t::MINUTES): {
 									// Получаем начало дня недели
 									const uint64_t begin = this->begin(date, type);
 									// Определяем количество прошедших минут
-									result = static_cast <uint64_t> ((static_cast <uint64_t> (86400000) - static_cast <uint64_t> (::floor(date - begin))) / 60000.);
+									result = static_cast <uint64_t> ((static_cast <uint64_t> (86400000) - static_cast <uint64_t> (::floor(static_cast <long double> (date - begin)))) / 60000.);
 								} break;
 								// Если нам нужно получить количество оставшихся секунд
 								case static_cast <uint8_t> (type_t::SECONDS): {
 									// Получаем начало дня недели
 									const uint64_t begin = this->begin(date, type);
 									// Определяем количество прошедших секунд
-									result = static_cast <uint64_t> ((static_cast <uint64_t> (86400000) - static_cast <uint64_t> (::floor(date - begin))) / 1000.);
+									result = static_cast <uint64_t> ((static_cast <uint64_t> (86400000) - static_cast <uint64_t> (::floor(static_cast <long double> (date - begin)))) / 1000.);
 								} break;
 								// Если нам нужно получить количество оставшихся миллисекунд
 								case static_cast <uint8_t> (type_t::MILLISECONDS): {
 									// Получаем начало дня недели
 									const uint64_t begin = this->begin(date, type);
 									// Определяем количество прошедших миллисекунд
-									result = static_cast <uint64_t> (static_cast <uint64_t> (86400000) - static_cast <uint64_t> (::floor(date - begin)));
+									result = static_cast <uint64_t> (static_cast <uint64_t> (86400000) - static_cast <uint64_t> (::floor(static_cast <long double> (date - begin))));
 								} break;
 								// Если нам нужно получить количество оставшихся микросекунд
 								case static_cast <uint8_t> (type_t::MICROSECONDS): {
@@ -2038,21 +2038,21 @@ uint64_t awh::Chrono::actual(const uint64_t date, const type_t value, const type
 									// Получаем начало дня недели
 									const uint64_t begin = this->begin(date, type);
 									// Определяем количество прошедших минут
-									result = static_cast <uint64_t> ((static_cast <uint64_t> (3600000) - static_cast <uint64_t> (::floor(date - begin))) / 60000.);
+									result = static_cast <uint64_t> ((static_cast <uint64_t> (3600000) - static_cast <uint64_t> (::floor(static_cast <long double> (date - begin)))) / 60000.);
 								} break;
 								// Если нам нужно получить количество оставшихся секунд
 								case static_cast <uint8_t> (type_t::SECONDS): {
 									// Получаем начало дня недели
 									const uint64_t begin = this->begin(date, type);
 									// Определяем количество прошедших секунд
-									result = static_cast <uint64_t> ((static_cast <uint64_t> (3600000) - static_cast <uint64_t> (::floor(date - begin))) / 1000.);
+									result = static_cast <uint64_t> ((static_cast <uint64_t> (3600000) - static_cast <uint64_t> (::floor(static_cast <long double> (date - begin)))) / 1000.);
 								} break;
 								// Если нам нужно получить количество оставшихся миллисекунд
 								case static_cast <uint8_t> (type_t::MILLISECONDS): {
 									// Получаем начало дня недели
 									const uint64_t begin = this->begin(date, type);
 									// Определяем количество прошедших миллисекунд
-									result = static_cast <uint64_t> (static_cast <uint64_t> (3600000) - static_cast <uint64_t> (::floor(date - begin)));
+									result = static_cast <uint64_t> (static_cast <uint64_t> (3600000) - static_cast <uint64_t> (::floor(static_cast <long double> (date - begin))));
 								} break;
 								// Если нам нужно получить количество оставшихся микросекунд
 								case static_cast <uint8_t> (type_t::MICROSECONDS): {
@@ -2079,14 +2079,14 @@ uint64_t awh::Chrono::actual(const uint64_t date, const type_t value, const type
 									// Получаем начало дня недели
 									const uint64_t begin = this->begin(date, type);
 									// Определяем количество прошедших секунд
-									result = static_cast <uint64_t> ((static_cast <uint64_t> (60000) - static_cast <uint64_t> (::floor(date - begin))) / 1000.);
+									result = static_cast <uint64_t> ((static_cast <uint64_t> (60000) - static_cast <uint64_t> (::floor(static_cast <long double> (date - begin)))) / 1000.);
 								} break;
 								// Если нам нужно получить количество оставшихся миллисекунд
 								case static_cast <uint8_t> (type_t::MILLISECONDS): {
 									// Получаем начало дня недели
 									const uint64_t begin = this->begin(date, type);
 									// Определяем количество прошедших миллисекунд
-									result = static_cast <uint64_t> (static_cast <uint64_t> (60000) - static_cast <uint64_t> (::floor(date - begin)));
+									result = static_cast <uint64_t> (static_cast <uint64_t> (60000) - static_cast <uint64_t> (::floor(static_cast <long double> (date - begin))));
 								} break;
 								// Если нам нужно получить количество оставшихся микросекунд
 								case static_cast <uint8_t> (type_t::MICROSECONDS): {
@@ -2113,7 +2113,7 @@ uint64_t awh::Chrono::actual(const uint64_t date, const type_t value, const type
 									// Получаем начало дня недели
 									const uint64_t begin = this->begin(date, type);
 									// Определяем количество прошедших миллисекунд
-									result = static_cast <uint64_t> (static_cast <uint64_t> (1000) - static_cast <uint64_t> (::floor(date - begin)));
+									result = static_cast <uint64_t> (static_cast <uint64_t> (1000) - static_cast <uint64_t> (::floor(static_cast <long double> (date - begin))));
 								} break;
 								// Если нам нужно получить количество оставшихся микросекунд
 								case static_cast <uint8_t> (type_t::MICROSECONDS): {
@@ -2733,7 +2733,7 @@ uint64_t awh::Chrono::actual(const uint64_t date, const type_t value, const type
 									// Получаем начало дня недели
 									const uint64_t begin = this->begin(date, type);
 									// Определяем количество прошедших секунд
-									result = static_cast <uint64_t> (::floor(date - begin));
+									result = static_cast <uint64_t> (::floor(static_cast <long double> (date - begin)));
 								} break;
 								// Если нам нужно получить количество прошедших микросекунд
 								case static_cast <uint8_t> (type_t::MICROSECONDS): {
@@ -2781,7 +2781,7 @@ uint64_t awh::Chrono::actual(const uint64_t date, const type_t value, const type
 									// Получаем начало дня недели
 									const uint64_t begin = this->begin(date, type);
 									// Определяем количество прошедших секунд
-									result = static_cast <uint64_t> (::floor(date - begin));
+									result = static_cast <uint64_t> (::floor(static_cast <long double> (date - begin)));
 								} break;
 								// Если нам нужно получить количество прошедших микросекунд
 								case static_cast <uint8_t> (type_t::MICROSECONDS): {
@@ -2822,7 +2822,7 @@ uint64_t awh::Chrono::actual(const uint64_t date, const type_t value, const type
 									// Получаем начало дня недели
 									const uint64_t begin = this->begin(date, type);
 									// Определяем количество прошедших секунд
-									result = static_cast <uint64_t> (::floor(date - begin));
+									result = static_cast <uint64_t> (::floor(static_cast <long double> (date - begin)));
 								} break;
 								// Если нам нужно получить количество прошедших микросекунд
 								case static_cast <uint8_t> (type_t::MICROSECONDS): {
@@ -2856,7 +2856,7 @@ uint64_t awh::Chrono::actual(const uint64_t date, const type_t value, const type
 									// Получаем начало дня недели
 									const uint64_t begin = this->begin(date, type);
 									// Определяем количество прошедших секунд
-									result = static_cast <uint64_t> (::floor(date - begin));
+									result = static_cast <uint64_t> (::floor(static_cast <long double> (date - begin)));
 								} break;
 								// Если нам нужно получить количество прошедших микросекунд
 								case static_cast <uint8_t> (type_t::MICROSECONDS): {
@@ -2883,7 +2883,7 @@ uint64_t awh::Chrono::actual(const uint64_t date, const type_t value, const type
 									// Получаем начало дня недели
 									const uint64_t begin = this->begin(date, type);
 									// Определяем количество прошедших секунд
-									result = static_cast <uint64_t> (::floor(date - begin));
+									result = static_cast <uint64_t> (::floor(static_cast <long double> (date - begin)));
 								} break;
 								// Если нам нужно получить количество прошедших микросекунд
 								case static_cast <uint8_t> (type_t::MICROSECONDS): {
